grep: Moved examples solution to example.rs
